feat :: 네비게이션

+++ app/src/main/java/team/retum/jobisandroidv2/root/RootNavigation.kt
@@ -18,7 +18,7 @@ fun NavGraphBuilder.root(
     onChangePasswordClick: () -> Unit,
     onReportBugClick: () -> Unit,
     navigateToLanding: () -> Unit,
-    onPostReviewClick: () -> Unit,
+    onPostReviewClick: (Long) -> Unit,
 ) {
     composable(NAVIGATION_ROOT) {
         Root(
